requirements.txt
altgraph==0.17.4
bcrypt==4.2.0
certifi==2024.7.4
cffi==1.16.0
cryptography==43.0.0
greenlet==3.0.3
influxdb-client==1.44.0
packaging==24.1
paho-mqtt==2.1.0
pefile==2023.2.7
pycparser==2.22
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32-ctypes==0.2.2
reactivex==4.0.4
setuptools==72.1.0
six==1.16.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
urllib3==2.2.2
wheel==0.44.0
